refactor server start
