+++ main.go
@@ -3,6 +3,6 @@ package main
 import "github.com/zywaited/delay-queue/server"
 
 func main() {
-	srv := server.NewDelayQueue()
+	srv := server.NewDelayQueue(server.AcquireServerInits()...)
 	panic(srv.Run())
 }
